+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.15 - not yet released
 * New features:
+  * Loading images on web.
 * Fixed bugs:
 * Infrastructure maintenance:
   * flutter pub upgrade.